+++ packages/cli/doc-plugin-medium-zoom/src/modern-app-env.d.ts
@@ -1,0 +1,2 @@
+/// <reference types='@modern-js/module-tools/types' />
+/// <reference types='@modern-js/plugin-testing/types' />