+++ packages/solutions/module-tools/src/utils/tspathsTransform.ts
@@ -1,10 +1,10 @@
 import * as path from 'path';
 import { createMatchPath } from '@modern-js/utils/tsconfig-paths';
 import { fs } from '@modern-js/utils';
-import * as parser from '../../compiled/@babel/parser';
-import generator from '../../compiled/@babel/generator';
-import * as t from '../../compiled/@babel/types';
-import traverse, { NodePath } from '../../compiled/@babel/traverse';
+import * as parser from '@babel/parser';
+import generator from '@babel/generator';
+import * as t from '@babel/types';
+import traverse, { NodePath } from '@babel/traverse';
 import { dtsAliasExts } from '../constants/file';
 // import { defaultTransformedFunctions } from '../constants/dts';
 // import { matchesPattern, isImportCall } from './dts';