+++ tests/e2e/builder/cases/css/css-modules-dom/tests/index.test.ts
@@ -1,0 +1,93 @@
+import { join, resolve } from 'path';
+import { fs } from '@modern-js/utils';
+import { build, getHrefByEntryName } from '@scripts/shared';
+import { webpackOnlyTest } from '@scripts/helper';
+import { expect } from '@modern-js/e2e/playwright';
+
+const fixtures = resolve(__dirname, '../');
+
+webpackOnlyTest('enableCssModuleTSDeclaration', async () => {
+  fs.removeSync(join(fixtures, 'src/App.module.less.d.ts'));
+  fs.removeSync(join(fixtures, 'src/App.module.scss.d.ts'));
+  const buildOpts = {
+    cwd: fixtures,
+    entry: {
+      main: join(fixtures, 'src/index.ts'),
+    },
+  };
+
+  const builder = await build(
+    buildOpts,
+    {
+      output: {
+        enableCssModuleTSDeclaration: true,
+      },
+    },
+    false,
+  );
+
+  expect(
+    fs.existsSync(join(fixtures, 'src/App.module.less.d.ts')),
+  ).toBeTruthy();
+
+  expect(
+    fs
+      .readFileSync(join(fixtures, 'src/App.module.less.d.ts'), {
+        encoding: 'utf-8',
+      })
+      .includes(`'title': string;`),
+  ).toBeTruthy();
+
+  expect(
+    fs.existsSync(join(fixtures, 'src/App.module.scss.d.ts')),
+  ).toBeTruthy();
+
+  expect(
+    fs
+      .readFileSync(join(fixtures, 'src/App.module.scss.d.ts'), {
+        encoding: 'utf-8',
+      })
+      .includes(`'header': string;`),
+  ).toBeTruthy();
+
+  builder.close();
+});
+
+webpackOnlyTest('disableCssExtract', async ({ page }) => {
+  const buildOpts = {
+    cwd: fixtures,
+    entry: {
+      main: join(fixtures, 'src/index.ts'),
+    },
+  };
+
+  const builder = await build(buildOpts, {
+    output: {
+      disableCssExtract: true,
+    },
+  });
+
+  await page.goto(getHrefByEntryName('main', builder.port));
+
+  // disableCssExtract worked
+  const files = await builder.unwrapOutputJSON();
+  const cssFiles = Object.keys(files).filter(file => file.endsWith('.css'));
+
+  expect(cssFiles.length).toBe(0);
+
+  // scss worked
+  await expect(
+    page.evaluate(
+      `window.getComputedStyle(document.getElementById('header')).fontSize`,
+    ),
+  ).resolves.toBe('20px');
+
+  // less worked
+  await expect(
+    page.evaluate(
+      `window.getComputedStyle(document.getElementById('title')).fontSize`,
+    ),
+  ).resolves.toBe('20px');
+
+  builder.close();
+});