+++ tests/integration/css/test/css.test.js
@@ -34,25 +34,6 @@ describe('default less loader options', () => {
 });
 
 describe('test css support', () => {
-  describe('base css support', () => {
-    it(`should emitted single css file `, async () => {
-      const appDir = path.resolve(fixtures, 'single-css');
-
-      await modernBuild(appDir);
-
-      const files = getCssFiles(appDir);
-
-      expect(files.length).toBe(1);
-
-      expect(
-        readFileSync(
-          path.resolve(appDir, `dist/static/css/${files[0]}`),
-          'utf8',
-        ),
-      ).toContain('#base-css{color:red}');
-    });
-  });
-
   describe('multi page css', () => {
     it('should emitted multiple css files', async () => {
       const appDir = path.resolve(fixtures, 'multi-css');
@@ -209,36 +190,6 @@ describe('test css support', () => {
     });
   });
 
-  describe('css url function', () => {
-    // FIXME: import svg in css files
-    // it('should use data-uri load image', async () => {
-    //   const appDir = path.resolve(fixtures, 'inline-css-url');
-
-    //   await modernBuild(appDir, [], { stderr: true });
-
-    //   const cssFiles = getCssFiles(appDir);
-
-    //   expect(cssFiles.length).toBe(1);
-
-    //   expect(readCssFile(appDir, cssFiles[0])).toMatch(
-    //     /background:url\(data:image\/svg\+xml;base64/,
-    //   );
-    // });
-    it('should keep image url', async () => {
-      const appDir = path.resolve(fixtures, 'keep-css-url');
-
-      await modernBuild(appDir);
-
-      const cssFiles = getCssFiles(appDir);
-
-      expect(cssFiles.length).toBe(1);
-
-      expect(readCssFile(appDir, cssFiles[0])).toMatch(
-        /background:url\(\/static\/image\/logo\.[a-z0-9]+\.png/,
-      );
-    });
-  });
-
   describe('css source map', () => {
     const getCssMaps = appDir =>
       readdirSync(path.resolve(appDir, 'dist/static/css')).filter(filepath =>
@@ -265,42 +216,11 @@ describe('test css support', () => {
 
       expect(cssMaps.length).toBe(0);
     });
-    // Todo skip, wait fix
-    it.skip(`should generate css ts declaration file`, async () => {
-      const appDir = path.resolve(fixtures, 'css-ts-declaration');
-
-      await modernBuild(appDir);
-
-      const generatedDTSFile = path.resolve(
-        appDir,
-        'src/index.module.css.d.ts',
-      );
-
-      expect(fs.readFileSync(generatedDTSFile, 'utf8')).toContain(
-        'export default cssExports',
-      );
-
-      fs.unlinkSync(generatedDTSFile);
-    });
   });
 });
 
 describe('less-support', () => {
   describe('base less support', () => {
-    it(`should emitted single css file`, async () => {
-      const appDir = resolve(fixtures, 'single-less');
-
-      await modernBuild(appDir);
-
-      const cssFiles = getCssFiles(appDir);
-
-      expect(cssFiles.length).toBe(1);
-
-      expect(readCssFile(appDir, cssFiles[0])).toContain(
-        '#header{height:20px;width:10px}',
-      );
-    });
-
     it(`should emitted multi css file`, async () => {
       const appDir = resolve(fixtures, 'multi-less');
 
@@ -327,20 +247,6 @@ describe('less-support', () => {
   });
 
   describe('base sass support', () => {
-    it(`should emitted single css file`, async () => {
-      const appDir = resolve(fixtures, 'single-sass');
-
-      await modernBuild(appDir);
-
-      const cssFiles = getCssFiles(appDir);
-
-      expect(cssFiles.length).toBe(1);
-
-      expect(readCssFile(appDir, cssFiles[0])).toContain(
-        '#header{height:20px;width:10px}',
-      );
-    });
-
     it(`should emitted multi css file`, async () => {
       const appDir = resolve(fixtures, 'multi-sass');